TelnetStream fixed, removing "workaround" call to .read() function
"About the TelnetStream read. I think this might be a bug in TelnetStream. The author may not have noticed though as their examples always read first anyway (for demonstration purposes). I will try and contact them to let them know..."

Following on from the above comment, I've been in touch with the author of TelnetStream and, while initially thinking it related to https://github.com/espressif/arduino-esp32/issues/4399 he has since discovered a fix. This is available now in 1,2,1 of the TelnetStream library which I have just tested and things now work without the .read()

So I'm removing it.